requirements.txt
# ============================================
# REQUIREMENTS.TXT PARA RENDER.COM
# ============================================
# Este archivo contiene todas las dependencias de producción del backend
# para compatibilidad con el build de Render.com
#
# Generado desde: backend/requirements/base.txt + backend/requirements/prod.txt
# ============================================

# ============================================
# CORE FRAMEWORK
# ============================================
# ✅ ACTUALIZADO 2025-01-27: fastapi 0.128.0 incluye starlette>=0.49.1 (corrige CVE-2025-54121, CVE-2025-62727)
# FastAPI 0.128.0 requiere starlette<0.51.0,>=0.40.0 - starlette 0.50.0 instalado automáticamente
fastapi>=0.128.0
uvicorn[standard]==0.38.0
python-multipart>=0.0.18
python-dotenv==1.2.1

# ============================================
# DATABASE
# ============================================
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
alembic==1.17.1

# ============================================
# VALIDATION AND SERIALIZATION
# ============================================
pydantic==2.12.4
pydantic-settings==2.12.0
email-validator==2.2.0
typing-extensions==4.15.0
annotated-types==0.7.0

# ============================================
# SECURITY
# ============================================
# ✅ MEJORA 2025-01-27: python-jose removido - NO se usa en el código
# El código usa PyJWT directamente con algoritmo HS256 (HMAC-SHA256), que NO requiere ECDSA
# Esto elimina completamente la dependencia de ecdsa y su vulnerabilidad CVE-2024-23342
PyJWT==2.8.0  # Usado directamente en app/core/security.py con algoritmo HS256 (no requiere ECDSA)
# ✅ COMPATIBILIDAD: passlib 1.7.4 lee bcrypt.__about__.__version__
# bcrypt 4.1.0+ removió __about__ → AttributeError en Render. Usar 4.0.x.
passlib[bcrypt]==1.7.4
bcrypt>=3.2.0,<4.1  # 4.0.x tiene __about__; 4.1+ rompe passlib
cryptography>=41.0.0  # Para encriptación de API Keys con Fernet (no requiere ecdsa para operaciones básicas)

# ============================================
# DATE AND TIME
# ============================================
python-dateutil==2.9.0.post0
pytz==2024.1

# ============================================
# HTTP CLIENT
# ============================================
# ✅ ACTUALIZADO 2025-01-27: httpx 0.28.1 compatible con aiohttp>=3.13.3 (corrige 8 CVEs)
httpx==0.28.1
requests==2.32.5
# ✅ ACTUALIZADO 2025-01-27: urllib3>=2.6.3 corrige 5 CVEs (CVE-2025-50182, CVE-2025-50181, CVE-2025-66418, CVE-2025-66471, CVE-2026-21441)
urllib3>=2.6.3

# ============================================
# FILE PROCESSING
# ============================================
openpyxl==3.1.5
pandas==2.3.3
numpy==2.3.4
scikit-learn==1.6.1
xgboost==2.1.3
reportlab==4.2.5

# ============================================
# NOTIFICATIONS
# ============================================
# twilio==8.2.0  # NO USADO - WhatsApp se integra con Meta Developers API
aiosmtplib==3.0.2
jinja2>=3.1.6

# ============================================
# TASK SCHEDULING
# ============================================
apscheduler==3.11.1

# ============================================
# MONITORING AND LOGGING
# ============================================
python-json-logger==2.0.7

# ============================================
# RATE LIMITING
# ============================================
slowapi==0.1.9

# ============================================
# CACHE (REDIS)
# ============================================
# ✅ MEJORA 2025-01-27: Redis para cache distribuido en producción
# Opcional: se usa MemoryCache como fallback si Redis no está disponible
redis>=5.0.0,<6.0.0

# ============================================
# MONITORING (SENTRY)
# ============================================
# ✅ MEJORA 2025-01-27: Sentry para monitoreo de errores en producción
# Opcional: solo se inicializa si SENTRY_DSN está configurado
sentry-sdk[fastapi]>=2.0.0

# ============================================
# GOOGLE (Drive, Sheets, Vision OCR - informe papeletas)
# ============================================
google-api-python-client>=2.100.0
google-auth>=2.25.0
google-cloud-vision>=3.5.0

# ============================================
# PRODUCTION SERVER
# ============================================
gunicorn==23.0.0
